student detail 수정

+++ src/main/java/com/green/smartgradever2/student/StudentRepository.java
@@ -3,7 +3,7 @@ package com.green.smartgradever2.student;
 import com.green.smartgradever2.config.entity.StudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<StudentEntity,Long> {
+public interface StudentRepository extends JpaRepository<StudentEntity,Integer> {
     StudentEntity findBystudentNum(int studentNum);
     StudentEntity findByStudentNum(Integer studentNum);
 }
