+++ src/main/java/com/green/smartgradever2/student/model/StudentLectureReg.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class StudentLectureReg {
     private Long ilecture;
-    private Long studentNum;
+    private Integer studentNum;
 }
